chore: rename lib/common/stddeque.h => lib/common/StdDeque.h
