+++ src/lib/base/LogOutputters.h
@@ -10,7 +10,7 @@
 
 #include "base/ILogOutputter.h"
 #include "common/Common.h"
-#include "common/stddeque.h"
+#include "common/StdDeque.h"
 #include "mt/Thread.h"
 
 #include <fstream>
