+++ src/lib/common/CMakeLists.txt
@@ -6,7 +6,7 @@ configure_file(Constants.h.in Constants.h @ONLY)
 add_library(common STATIC
   Common.h
   IInterface.h
-  stddeque.h
+  StdDeque.h
   stdexcept.h
   stdistream.h
   stdlist.h
